web/backend/requirements.txt
fastapi==0.115.0
uvicorn==0.24.0
sqlalchemy==2.0.20
pydantic==2.0.3
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1
python-multipart==0.0.6
python-dotenv==1.0.0
pydantic-settings==2.0.3
email-validator==2.1.0
httpx==0.26.0
# 数据库驱动（根据需要启用）
pymysql==1.1.0
psycopg2-binary==2.9.9 